keystone user provider fix for v3 calls
When the keystone user provider does it's prefetch for
the user list, it is not passing the --domain parameter
in the call. This results in the cli receiving a 401
error and puts the openstack provider in an error state
that causes all other providers dependent on the openstack
provider to error out.

This change adds the the --domain parameter with the default
domain as the value.

Change-Id: I3bf61cc460a0c1ee3dffdf1da14d85c3563fb772
Closes-Bug: 1554555